Fair ametsuchi skeleton
