+++ irohad/ametsuchi/CMakeLists.txt
@@ -1,8 +1,13 @@
 add_library(ametsuchi
+    impl/flat_file/flat_file.cpp
     impl/storage_impl.cpp
+    impl/temporary_wsv_impl.cpp
+    impl/mutable_storage_impl.cpp
     )
 
 target_link_libraries(ametsuchi
     rxcpp
     optional
+    pqxx
+    cpp_redis
     )
