--- irohad/ametsuchi/block_store/CMakeLists.txt
@@ -1,3 +1,0 @@
-add_library(block_store
-  backend/flat_file.cpp
-  )
